refactor: 디자인 및 코드 정리
