+++ src/App.vue
@@ -8,25 +8,14 @@
         'with-navbar': !$route.meta.hideNavBar,
       }"
     >
-      <KeepAlive>
-        <router-view />
-      </KeepAlive>
+      <router-view />
     </div>
     <NavBar v-if="!$route.meta.hideNavBar" />
   </div>
 </template>
 
-<script>
-import AppBar from '@/components/AppBar.vue';
-import NavBar from '@/components/NavBar.vue';
-
-export default {
-  name: 'App',
-  components: {
-    AppBar,
-    NavBar,
-  },
-};
+<script setup>
+import { NavBar, AppBar } from '@/components';
 </script>
 
 <style lang="scss">
@@ -35,7 +24,6 @@ export default {
   display: flex;
   flex-direction: column;
   background-color: var(--color-background);
-  font-family: var(--font-regular);
 }
 
 .content {
