+++ src/views/Auth/FindPasswordComplete.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script setup lang="ts">
-import Button from '@/components/button/Button.vue';
+import { Button } from '@/components';
 import { useRouter } from 'vue-router';
 
 const router = useRouter();
